Allow Elixir language in dialyzer sensor
Dialyzer works on .beam files, so the Elixir library dialyxir
can produce a file with the same format and same messages.

The only difference is that in Elixir multiple files can have the
same filename if they are in different directories, so I brought
back the old match, which tried to find the file using the full
path first. If it fails, then it falls back to the currently
working solution.